Computation of numerical mean-field heat capacity for TmVO4-RF-E. Computed numerical Cp at all multiples of 1000 Oe from 1 to 8 using the magnetic field distribution computed by COMSOL for the 'basic' half-diamond shape as modeled yesterday. Results are satisfactory. Will send to Ian. Comsol has issues computing the distribution at all fields and all temperatures in one go, so the computation had to be broken down into several steps and one MFD had to be exported for each step. As a result, the main Matlab script was adapted so as to allow for combining several files, each containing a variable number of MFDs, sometimes at different values of temperature.
Created a function that allows to save the results of the computation into .dat files, so as to avoid having to re-run the computation every time.

Turned off output of fsolve in 'OP_TFIM_array.m', which is useless when using fsolve many times.
This turned out to reduce the computation time by a factor 7!

Also disabled the warning in 'FSCp_TFIM.m', which was also output each time the function was invoked, i.e. too many to be useful.
(That did not significantly reduce the computation time however.)

Added a note in 'tools.m' about how to run a for loop containing (sub)sections.